Allow QDial to draw its hover circle outside its bounds
Added a QFocusFrame, which allows for widgets to be drawn outside o their bounds, to
QDial. This is used to draw the hover circle. Not having to fit the hover circle inside
the widget allows both the hover circle and the dial to be larger.

QFocusRect unfortunately doesn't update on all operations (e.g. drag or scroll),
there is a workaround, but it may impact performarce.