Allow QSlider to draw its hover circle outside its bounds
Added a FocusFrame, which allows for widgets to be drawn outside of their bounds.
This is used to draw the hover circle. Not having to fit the hover circle inside
the widget allows the slider to take less space.

QFocusFrame unfortunately doesn't update on all operations (e.g. drag or scroll),
there is a workaround: SliderFocusFrame which is a custom class inheriting from QFocusFrame,
which also captures paint events from its focused widget.

Also converted QDial to also use SliderFocusFrame