Make the tab implementation compatible with QCommonStyle
Make CE_TabBarTabLabel and SE_TabBarTabText compatible with QCommon style
(i.e. either of theese implementation may now be replaced with the one from
QCommonStyle and there should not be any major issues)

This fixes a bug wich caused the tab labels to not elide correctly,
because the eliding code in Qt was expecting different layout then
previously used.

If an app draws custom elements using the style, the result should be now
more predictable.

This implementation also has a benefit of being simpler.